+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
+**3.3.4 - 02/19/25**
+
+  - Type-hinting: Fix mypy errors in tests/framework/lookup/test_lookup.py
+
+**3.3.3 - 02/14/25**
+
+  - Type-hinting: Fix mypy errors in tests/framework/components/test_manager.py
+
+**3.3.2 - 02/12/25**
+
+  - Type-hinting: Fix mypy errors in tests/framework/components/test_parser.py
+  - Type-hinting: Fix mypy errors in tests/framework/components/test_component.py
+
 **3.3.1 - 02/07/25**
 
   - Type-hinting: Fix mypy errors in tests/framework/lookup/test_interpolation.py